Added Console Command
The command copied is meant to be placed into the browsers console. The code will only work on sparxmaths website. If you leave or refresh the page, the JS snippet must be entered into the console again to resume effect.

The JS snippet adds additional scripts to the sparxmaths website, which enhances the User Experience by adding Additional features listed in the README.md